chore: locale slug
