+++ src/app/[locale]/components/Body/Main.tsx
