+++ src/app/[locale]/components/Body/Root.tsx
