+++ src/app/[locale]/components/Body/index.ts
