+++ src/app/[locale]/components/Buttons/SessionInfo.tsx
