+++ src/app/[locale]/components/Menu/index.ts
